requirement.txt
beautifulsoup4==4.9.3
googlesearch-python==1.1.0
requests==2.25.1